Deprecate isEqual, isLess, isLessThanOrEqualTo
These are occasionally useful as hooks, but should really be removed to match the way that the Equatable/Comparable operations are implemented on the integer protocols, as this makes the stdlib more consistent and predictable.

These were originally used because we didn't have static operators at the time; they had to be free functions that called these methods as implementations; that hasn't been necessary since Swift 4, but somehow this machinery never got cleaned up.