Make out-of-range integer literals trap in generic contexts instead of silently returning zero.
As a temporary hack, this just fatalErrors(), because almost everything in the standard lib depends on integer literals, so you can't do much else without introducing an infinite loop of @_transparent functions (there wouldn't really be an infinite loop, but we can't see that at the SIL level). Obviously, we'll need to get a diagnostic in here sometime.

Also this still doesn't pass a bunch of tests that are looking for specific IR sequences at the SIL level, because there's now an additional branch (that gets optimized, but that's post-SIL).